Change the syntax of the generic identifier reference.

+++ slake/runtime.cc
@@ -295,11 +295,11 @@ SLAKE_API Value Runtime::defaultValueOf(const TypeRef &type) const {
 		case TypeId::Array:
 			return Value(slake::Reference(nullptr));
 		case TypeId::Ref:
-			return Value(slake::ObjectFieldRef(nullptr, SIZE_MAX));
+			return Value(slake::ObjectFieldRef(nullptr, UINT32_MAX));
 			break;
 		case TypeId::TempRef:
 			// TODO: Replace it with TempRef version.
-			return Value(slake::ObjectFieldRef(nullptr, SIZE_MAX));
+			return Value(slake::ObjectFieldRef(nullptr, UINT32_MAX));
 			break;
 		default:
 			break;
